redis - feat: moving create helper functions to create.ts
